+++ BloodSplatterScript.cs
@@ -17,7 +17,7 @@ public class BloodSplatterScript : MonoBehaviour
     {
         if (Input.GetMouseButtonDown(0))
         {
-            this.bloodRotation.Rotate((float) 0f, (float) this.bloodLocalRotationYOffset, (float) 0f);
+            this.bloodRotation.Rotate(0f, this.bloodLocalRotationYOffset, 0f);
             Transform transform1 = UnityEngine.Object.Instantiate(this.bloodPrefab, this.bloodPosition.position, this.bloodRotation.rotation) as Transform;
             this.bloodInstances = GameObject.FindGameObjectsWithTag("blood");
             if (this.bloodInstances.Length >= this.maxAmountBloodPrefabs)
