+++ ColorAffector.cs
@@ -62,7 +62,7 @@ public class ColorAffector : Affector
                     index--;
                 }
                 int num3 = index + 1;
-                float num4 = this.GradualLen / (float) (this.ColorArr.Length - 1);
+                float num4 = this.GradualLen / (this.ColorArr.Length - 1);
                 float t = (this.ElapsedTime - num4 * index) / num4;
                 base.Node.Color = Color.Lerp(this.ColorArr[index], this.ColorArr[num3], t);
             }
