+++ PanWithMouse.cs
@@ -25,8 +25,8 @@ public class PanWithMouse : IgnoreTimeScale
         {
             this.range = 0.1f;
         }
-        float x = Mathf.Clamp((float) ((mousePosition.x - num2) / num2 / this.range), (float) -1f, (float) 1f);
-        float y = Mathf.Clamp((float) ((mousePosition.y - num3) / num3 / this.range), (float) -1f, (float) 1f);
+        float x = Mathf.Clamp((mousePosition.x - num2) / num2 / this.range, -1f, 1f);
+        float y = Mathf.Clamp((mousePosition.y - num3) / num3 / this.range, -1f, 1f);
         this.mRot = Vector2.Lerp(this.mRot, new Vector2(x, y), num * 5f);
         this.mTrans.localRotation = this.mStart * Quaternion.Euler(-this.mRot.y * this.degrees.y, this.mRot.x * this.degrees.x, 0f);
     }
