+++ PopuplistCharacterSelection.cs
@@ -27,10 +27,10 @@ public class PopuplistCharacterSelection : MonoBehaviour
                 stat = costume.stat;
             }
         }
-        this.SPD.transform.localScale = new Vector3((float) stat.SPD, 20f, 0f);
-        this.GAS.transform.localScale = new Vector3((float) stat.GAS, 20f, 0f);
-        this.BLA.transform.localScale = new Vector3((float) stat.BLA, 20f, 0f);
-        this.ACL.transform.localScale = new Vector3((float) stat.ACL, 20f, 0f);
+        this.SPD.transform.localScale = new Vector3(stat.SPD, 20f, 0f);
+        this.GAS.transform.localScale = new Vector3(stat.GAS, 20f, 0f);
+        this.BLA.transform.localScale = new Vector3(stat.BLA, 20f, 0f);
+        this.ACL.transform.localScale = new Vector3(stat.ACL, 20f, 0f);
     }
 }
