+++ TITAN_CONTROLLER.cs
@@ -75,7 +75,7 @@ public class TITAN_CONTROLLER : MonoBehaviour
             if (num2 != 0 || num != 0)
             {
                 y = this.currentCamera.transform.rotation.eulerAngles.y;
-                num4 = Mathf.Atan2((float) num, (float) num2) * 57.29578f;
+                num4 = Mathf.Atan2(num, num2) * 57.29578f;
                 num4 = -num4 + 90f;
                 num5 = y + num4;
                 this.targetDirection = num5;
@@ -130,7 +130,7 @@ public class TITAN_CONTROLLER : MonoBehaviour
             if (num2 != 0 || num != 0)
             {
                 y = this.currentCamera.transform.rotation.eulerAngles.y;
-                num4 = Mathf.Atan2((float) num, (float) num2) * 57.29578f;
+                num4 = Mathf.Atan2(num, num2) * 57.29578f;
                 num4 = -num4 + 90f;
                 num5 = y + num4;
                 this.targetDirection = num5;
