+++ TypewriterEffect.cs
@@ -16,14 +16,14 @@ public class TypewriterEffect : MonoBehaviour
             this.mLabel = base.GetComponent<UILabel>();
             this.mLabel.supportEncoding = false;
             this.mLabel.symbolStyle = UIFont.SymbolStyle.None;
-            this.mText = this.mLabel.font.WrapText(this.mLabel.text, (float) this.mLabel.lineWidth / this.mLabel.cachedTransform.localScale.x, this.mLabel.maxLineCount, false, UIFont.SymbolStyle.None);
+            this.mText = this.mLabel.font.WrapText(this.mLabel.text, this.mLabel.lineWidth / this.mLabel.cachedTransform.localScale.x, this.mLabel.maxLineCount, false, UIFont.SymbolStyle.None);
         }
         if (this.mOffset < this.mText.Length)
         {
             if (this.mNextChar <= Time.time)
             {
                 this.charsPerSecond = Mathf.Max(1, this.charsPerSecond);
-                float num = 1f / (float) this.charsPerSecond;
+                float num = 1f / this.charsPerSecond;
                 switch (this.mText[this.mOffset])
                 {
                     case '.':
