+++ UICursor.cs
@@ -56,8 +56,8 @@ public class UICursor : MonoBehaviour
             Vector3 mousePosition = Input.mousePosition;
             if (this.uiCamera != null)
             {
-                mousePosition.x = Mathf.Clamp01(mousePosition.x / (float) Screen.width);
-                mousePosition.y = Mathf.Clamp01(mousePosition.y / (float) Screen.height);
+                mousePosition.x = Mathf.Clamp01(mousePosition.x / Screen.width);
+                mousePosition.y = Mathf.Clamp01(mousePosition.y / Screen.height);
                 this.mTrans.position = this.uiCamera.ViewportToWorldPoint(mousePosition);
                 if (this.uiCamera.isOrthoGraphic)
                 {
