+++ UIRoot.cs
@@ -66,15 +66,15 @@ public class UIRoot : MonoBehaviour
         height = Mathf.Max(2, height);
         if (this.scalingStyle == Scaling.FixedSize)
         {
-            return (float) this.manualHeight / (float) height;
+            return this.manualHeight / (float)height;
         }
         if (height < this.minimumHeight)
         {
-            return (float) this.minimumHeight / (float) height;
+            return this.minimumHeight / (float)height;
         }
         if (height > this.maximumHeight)
         {
-            return (float) this.maximumHeight / (float) height;
+            return this.maximumHeight / (float)height;
         }
         return 1f;
     }
@@ -114,7 +114,7 @@ public class UIRoot : MonoBehaviour
             {
                 float x = 2f / activeHeight;
                 Vector3 localScale = this.mTrans.localScale;
-                if (Mathf.Abs((float) (localScale.x - x)) > float.Epsilon || Mathf.Abs((float) (localScale.y - x)) > float.Epsilon || Mathf.Abs((float) (localScale.z - x)) > float.Epsilon)
+                if (Mathf.Abs(localScale.x - x) > float.Epsilon || Mathf.Abs(localScale.y - x) > float.Epsilon || Mathf.Abs(localScale.z - x) > float.Epsilon)
                 {
                     this.mTrans.localScale = new Vector3(x, x, x);
                 }
