+++ UIViewport.cs
@@ -15,7 +15,7 @@ public class UIViewport : MonoBehaviour
         {
             Vector3 vector = this.sourceCamera.WorldToScreenPoint(this.topLeft.position);
             Vector3 vector2 = this.sourceCamera.WorldToScreenPoint(this.bottomRight.position);
-            Rect rect = new Rect(vector.x / (float) Screen.width, vector2.y / (float) Screen.height, (vector2.x - vector.x) / (float) Screen.width, (vector.y - vector2.y) / (float) Screen.height);
+            Rect rect = new Rect(vector.x / Screen.width, vector2.y / Screen.height, (vector2.x - vector.x) / Screen.width, (vector.y - vector2.y) / Screen.height);
             float num = this.fullSize * rect.height;
             if (rect != this.mCam.rect)
             {
